ENHANCEMENT: Add ReflectedColor
WHY: Provide support to calculate reflected color.
HOW: Create new function called ReflectedColor that use the Reflect
vector in the material to determine the color reflected.
NOTE: This version always returns black for color. Although this is
wrong in the general case it is sufficient to make the test pass for
now.